Adds refresh functionality to commitlist
