+++ src/app/detailed-viewer/commit-list/commit-list.component.css
@@ -12,10 +12,13 @@
 
 .flex-container {
   display: flex;
-  justify-content: space-between;
   flex-direction: row;
   align-items: center;
   width: 100%;
+}
+
+.space-apart {
+  justify-content: space-between;
 }
 
 .hidden_scrollbar {
